Fix two digit positions (separated with '-')
The original syntax check was done in ParseChord(), but the regular
expression was broken (->fixed). Errors at this point lead to a big
error-cross instead of a chord symbol.

Before that there is a rough length of "positions" and "fingering"
test in GenerateChordHtml(). This test is extended (as proposed) to
allow more than 6 chars if there is a '-' in "positions".

Test-case is added to "test.html".

I personally don't like the checking to be done at two different
places...